Switch to new resolution system in ThemeRegistry
